+++ extensions/share-my-code/src/Constants.ts
@@ -1,0 +1,7 @@
+export const baseURL = "https://sharemycode.fr/ajax";
+export const smcUrl = "https://sharemycode.fr";
+
+export const MIN_SLUG_SIZE = 3;
+export const RAND_SLUG_SIZE = 4;
+
+export const characters = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";